refactor(frontend): Create folder interfaces in src
