+++ frontend/src/components/interface/compilingIndicator/CompilingIndicator.vue
