+++ frontend/src/components/interface/drawer/NavigationDrawer.vue
